+++ src/components/svelte/Button.svelte
@@ -1,8 +1,6 @@
 <script lang="ts">
   export let buttonText: string = "Click me";
-
-  import { Button } from 'flowbite-svelte';
-
+  import { Button } from "flowbite-svelte";
 </script>
 
 <Button>{buttonText}</Button>
