+++ src/pages/match/[matchId].astro
@@ -1,7 +1,7 @@
 ---
 export const prerender = false;
 
-import CenterLayout from "@layouts/CenterLayout.astro";
+import CenterLayout from "@layouts/Layout.astro";
 import App from "@components/svelte/App.svelte";
 
 const { matchId } = Astro.params;
